backend: move ATSInterface to its own file
